Changes 2.0
Added the original Author back after changes to older files.
Added the ability to get name description and price from the datafeed with the value show="name,description,price"
Added it to show up as a table to make it easyer to put on a website.
Added the ability to put in more then one pid like this pid="1,2,3,4" etc to have it fetch multiple ids and show them in a table.
Added the ability to translate the Table names via WordPress translation file.